divided MimeTypeChecker + test
